slider fixes
